Add max_size() adapter
This uses the given chunking method but also limits the chunk size to a
maximum value.

Note that the inner chunking method IS reset when a chunk boundary is
emitted because of the size limit. That means that using a size limit
will not only add new boundary, inside of blocks too big, it might cause
the boundary after such a one to not happen anymore.